requirements.txt
asgiref==3.8.1
black==25.1.0
cfgv==3.4.0
click==8.1.8
distlib==0.3.9
Django==5.2
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
filelock==3.18.0
identify==2.6.10
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
pathspec==0.12.1
pillow==11.2.1
platformdirs==4.3.7
pre_commit==4.2.0
PyJWT==2.9.0
PyYAML==6.0.2
sqlparse==0.5.3
virtualenv==20.30.0
